Созданы методы для обработки исключений, и отправке сообщения об этом клиенту. Обработчик протестирован через Postman( все методы работают корректно)
