+++ src/main/java/faang/school/postservice/client/UserServiceClient.java
@@ -12,7 +12,7 @@ import java.util.List;
 @FeignClient(name = "user-service", url = "${user-service.host}:${user-service.port}")
 public interface UserServiceClient {
 
-    @GetMapping("/user/{userId}")
+    @GetMapping("/users/{userId}")
     UserDto getUser(@PathVariable("userId") long userId);
 
     @PostMapping("/users")
