+++ src/main/java/faang/school/postservice/dto/client/UserDto.java
@@ -1,10 +1,14 @@
 package faang.school.postservice.dto.client;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class UserDto {
 
     private long id;
