--- src/main/java/faang/school/postservice/exceptions/DataAlreadyExistingException.java
@@ -1,7 +1,0 @@
-package faang.school.postservice.exceptions;
-
-public class DataAlreadyExistingException extends DataValidationException{
-    public DataAlreadyExistingException(String message) {
-        super(message);
-    }
-}
