--- src/main/java/faang/school/postservice/exceptions/DataNotFoundException.java
@@ -1,7 +1,0 @@
-package faang.school.postservice.exceptions;
-
-public class DataNotFoundException extends RuntimeException {
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-}
