--- src/main/java/faang/school/postservice/exceptions/DataValidationException.java
@@ -1,7 +1,0 @@
-package faang.school.postservice.exceptions;
-
-public class DataValidationException extends RuntimeException{
-    public DataValidationException(String message) {
-        super(message);
-    }
-}
