--- src/main/java/faang/school/postservice/exceptions/ErrorResponse.java
@@ -1,13 +1,0 @@
-package faang.school.postservice.exceptions;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-@AllArgsConstructor
-public class ErrorResponse {
-
-    private String message;
-}
