--- src/main/java/faang/school/postservice/exceptions/SameTimeActionException.java
@@ -1,7 +1,0 @@
-package faang.school.postservice.exceptions;
-
-public class SameTimeActionException extends RuntimeException{
-    public SameTimeActionException(String message) {
-        super(message);
-    }
-}
