+++ src/test/java/faang/school/postservice/service/LikeServiceTest.java
@@ -3,9 +3,9 @@ package faang.school.postservice.service;
 import faang.school.postservice.client.UserServiceClient;
 import faang.school.postservice.dto.LikeDto;
 import faang.school.postservice.dto.PostDto;
-import faang.school.postservice.exceptions.DataNotFoundException;
+import faang.school.postservice.exception.DataNotFoundException;
 import faang.school.postservice.dto.client.UserDto;
-import faang.school.postservice.exceptions.SameTimeActionException;
+import faang.school.postservice.exception.SameTimeActionException;
 import faang.school.postservice.mapper.LikeMapper;
 import faang.school.postservice.mapper.LikeMapperImpl;
 import faang.school.postservice.mapper.PostMapper;
